# openapi/baseline.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{slug}/webhooks": {
      "get": {
        "operationId": "WebhooksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      },
      "post": {
        "operationId": "WebhooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/v1/workspaces/{slug}/webhooks/{id}": {
      "get": {
        "operationId": "WebhooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      },
      "patch": {
        "operationId": "WebhooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      },
      "delete": {
        "operationId": "WebhooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
...
    "/api/v1/workspaces/{slug}/integrations/{provider}": {
      "delete": {
        "operationId": "IntegrationsController_disconnect",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Integrations"
        ]
      }
    },
...
    "/api/v1/workspaces/{slug}/projects/{projectSlug}/tasks/{number}/github-links": {
      "get": {
        "operationId": "GithubLinkController_list",
        "parameters": [
          {
            "name": "projectSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GithubLink"
        ]
      },
      "post": {
        "operationId": "GithubLinkController_create",
        "parameters": [
          {
            "name": "projectSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GithubLink"
        ]
      }
    },
    "/api/v1/workspaces/{slug}/projects/{projectSlug}/tasks/{number}/github-links/{linkId}": {
      "delete": {
        "operationId": "GithubLinkController_remove",
        "parameters": [
          {
            "name": "projectSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "GithubLink"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tasker Public API",
    "description": "Public REST API for the Tasker platform. All endpoints live under `/api/v1` and return Problem Details (RFC 7807) on error. Sensitive POSTs accept an `Idempotency-Key` header. See `/api/v1/docs` for the Swagger UI.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Current origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT or tsk_live_… API key",
        "type": "http",
        "description": "Two authentication mechanisms accept the same header shape: (1) the JWT used by the web UI, or (2) an API key minted from Settings → Platform → API keys."
      }
    },
    "schemas": {
...
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "maxLength": 500
          },
          "eventTypes": {
            "type": "array",
            "minItems": 1,
            "maxItems": 13,
            "items": {
              "type": "string",
              "enum": [
                "TASK_CREATED",
                "TASK_UPDATED",
                "TASK_DELETED",
                "PROJECT_CREATED",
                "PROJECT_UPDATED",
                "PROJECT_DELETED",
                "COMMENT_CREATED",
                "COMMENT_UPDATED",
                "COMMENT_DELETED",
                "SPRINT_CREATED",
                "SPRINT_UPDATED",
                "SPRINT_STARTED",
                "SPRINT_COMPLETED"
              ]
            }
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "url",
          "eventTypes"
        ]
      },
      "UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "maxLength": 500
          },
          "eventTypes": {
            "type": "array",
            "minItems": 1,
            "maxItems": 13,
            "items": {
              "type": "string",
              "enum": [
                "TASK_CREATED",
                "TASK_UPDATED",
                "TASK_DELETED",
                "PROJECT_CREATED",
                "PROJECT_UPDATED",
                "PROJECT_DELETED",
                "COMMENT_CREATED",
                "COMMENT_UPDATED",
                "COMMENT_DELETED",
                "SPRINT_CREATED",
                "SPRINT_UPDATED",
                "SPRINT_STARTED",
                "SPRINT_COMPLETED"
              ]
            }
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateTaskLinkDto": {
        "type": "object",
        "properties": {
          "externalRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "externalType": {
            "type": "string",
            "enum": [
              "ISSUE",
              "PR"
            ]
          }
        },
        "required": [
          "externalRef",
          "externalType"
        ]
      }
    }
  }
}